Now creates and opens an Xcode project.
Previously the open command would - create the package, update the Sources
folder, and open the package in Xcode.

Now instead of "open the package in Xcode" it does - create an Xcode
project from the package and open the Xcode project in Xcode.

My reasoning for this change is that Xcode handles the "New File"
command better when processing an Xcode project instead of a Swift
package.