Revamped the way the oldFile is used.
The make command used to copy the original version of dayxx.swift to
old-dayxx.swift and leave it there for reference.  Now, after the
package has been created, it moves old-dayxx.swift back to dayxx.swift.
It is still available for reference but the workflow is simplified in
the normal case.