Solved the infinite loop kluge.
Now instead of an explicit check for the "infinite loop" item I detect
the infinite loop and interrupt it.  I detect it by keeping track of the
maximum length of output from each command executed so far and comparing
the current output to that maximum.  If the current output is too big
then I terminate the loop.