2022 day19 part2, doesn't work.
The strategy for part1 is to find the paths that build the first geode
robot the fastest.  Starting from those paths find the paths that build
the second geode robot the fastest, etc.

This strategy doesn't work for part2 because the extra time allows you
to build more obsidian robots in the beginning which then allows you to
build geode robots faster towards the end.