2022 day19 added an alternate part1 solution.
The original part1 algorithm makes the first geode robot as fast as
possible, then from there makes the second geode robot as fast as
possible, etc.  This algorithm works for part1 but not part2.

The alternate part1 algorithm is the same as the part2 algorithm.  That
is just a depth first search with pruning of branches that can't
possibly make more geodes than the best already found.

Note that the original algorithm is faster on the test data and slower
on the real data.