Replaced the awkward patch mechanism.
Very rarely a problem solution requires access to an external library.
In this case the Package.swift file must be modified to specify the
reference to the external library.

To solve this problem I was using a combination of diff and patch to
avoid having to make the Package.swift changes manually each time I
wanted to run a problem solution that required an external library. This
method was prone to breaking anytime Apple changed SPM or Xcode.

Instead, I now just keep a copy of the custom Package.swift file in a
packages directory and use that each time a solution with an external
library is opened.